fix: styling for IndexBuilderFooter
